refactor: 파일 페이지 별로 정리
