+++ src/components/completePaymentsRegister/CompletePaymentsRegister.tsx
@@ -1,9 +1,9 @@
-import Button from './Button';
-import { COLOR } from '../styles/color';
-import { CardIssuer } from '../type';
-import ISSUER_KOREAN from '../constants/issuerKorean';
+import Button from '../Button';
+import { COLOR } from '../../styles/color';
+import { CardIssuer } from '../../type';
+import ISSUER_KOREAN from '../../constants/issuerKorean';
 import { Link } from 'react-router-dom';
-import PATH from '../constants/path';
+import PATH from '../../constants/path';
 import registerComplete from '../Images/register_complete.png';
 import styled from '@emotion/styled';
