+++ src/components/wrongAccess/WrongAccess.tsx
@@ -1,7 +1,7 @@
-import Button from './Button';
-import { COLOR } from '../styles/color';
+import Button from '../Button';
+import { COLOR } from '../../styles/color';
 import { Link } from 'react-router-dom';
-import PATH from '../constants/path';
+import PATH from '../../constants/path';
 import styled from '@emotion/styled';
 import wrongCat from '../Images/wrong_cat.png';
