+++ src/hooks/payments/useCardCVC.tsx
@@ -1,9 +1,9 @@
-import { isOnlyDigit, isSameLength } from '../domain/checkIsValid';
+import { isOnlyDigit, isSameLength } from '../../domain/checkIsValid';
 
-import { BOUND } from '../constants/number';
-import { ERROR_MESSAGE } from '../constants/message';
+import { BOUND } from '../../constants/number';
+import { ERROR_MESSAGE } from '../../constants/message';
 import { useState } from 'react';
-import useValidateInput from './useValidateInput';
+import useValidateInput from '../useValidateInput';
 
 const validateInputProps = {
   validatorPropsArray: [
@@ -13,20 +13,17 @@ const validateInputProps = {
     },
     {
       checkIsValid: (string: string) =>
-        isSameLength(string, BOUND.cardPasswordHeadUpper),
-      errorMessage: `${BOUND.cardPasswordHeadUpper}${ERROR_MESSAGE.invalidLengthTail}`,
+        isSameLength(string, BOUND.cardCVCUpper),
+      errorMessage: `${BOUND.cardCVCUpper}${ERROR_MESSAGE.invalidLengthTail}`,
     },
   ],
 };
 
-export default function useCardPasswordHead() {
+export default function useCardCVC() {
+  const [isTouched, setIsTouched] = useState(false);
   const validateInput = useValidateInput(validateInputProps);
 
-  const [isTouched, setIsTouched] = useState(false);
-
-  const isValid =
-    validateInput.inputValue.length === BOUND.cardPasswordHeadUpper &&
-    validateInput.errorMessage === '';
+  const isValid = isTouched && validateInput.errorMessage === '';
 
   return {
     cardCVC: validateInput.inputValue,
@@ -38,13 +35,13 @@ export default function useCardPasswordHead() {
     isValid,
     isTouched,
     initValue: () => {
-      setIsTouched(false);
       validateInput.initValue();
+      setIsTouched(false);
     },
   };
 }
 
-export interface UseCardPasswordHead {
+export interface UseCardCVC {
   cardCVC: string;
   onChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
   errorMessage: string;
