+++ src/hooks/payments/useCardNumbers.tsx
@@ -1,10 +1,10 @@
-import { isOnlyDigit, isSameLength } from '../domain/checkIsValid';
+import { isOnlyDigit, isSameLength } from '../../domain/checkIsValid';
 
-import { BOUND } from '../constants/number';
-import { ERROR_MESSAGE } from '../constants/message';
-import useLastValidValue from './useLastValidValue';
+import { BOUND } from '../../constants/number';
+import { ERROR_MESSAGE } from '../../constants/message';
+import useLastValidValue from '../useLastValidValue';
 import { useState } from 'react';
-import useValidateInput from './useValidateInput';
+import useValidateInput from '../useValidateInput';
 
 const cardNumberValidateInputProps = {
   validatorPropsArray: [
