+++ src/stories/CardPreview.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import CardPreview from '../components/CardPreview';
+import CardPreview from '../components/payments/CardPreview';
 
 const meta = {
   title: 'CardPreview',
